pyproject.toml
[project]
name = "foundryqs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=2.0.0b3",
    "azure-identity>=1.25.1",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
]
